feat: Move configurable behaviour to sharedkernel
