+++ Assets/Scripts/Blocks/BlockBehaviours/BehaviourBuilder.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
+using Assets.Scripts.SharedKernel;
 
 namespace Assets.Scripts.Blocks
 {
