+++ Assets/Scripts/Powerups.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a5aa490fa303d844ab94b47b624772a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
