+++ Assets/Scripts/Powerups/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f62d5df0c52fbbf4f9d3521993a7d8e1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
