+++ Assets/Scripts/Powerups/Interfaces/IPowerupEffect.cs
@@ -1,0 +1,7 @@
+namespace Assets.Scripts.Powerups
+{
+    public interface IPowerupEffect
+    {
+        void Apply();
+    }
+}
