+++ Assets/Scripts/Powerups/Interfaces/IPowerupEffect.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: df6cb0e5bdce4934cb0cbcfcee918a9f
