+++ Assets/Scripts/Powerups/PowerupType.cs
@@ -1,0 +1,14 @@
+namespace Assets.Scripts.Powerups
+{
+    public enum PowerupType
+    {
+        ExpandPaddle,
+        ShrinkPaddle,
+        ExtraLife,
+        MultiBall,
+        Laser,
+        Magnet,
+        SlowBall,
+        FastBall
+    }
+}
