+++ Assets/Scripts/Powerups/PowerupType.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 07f67db45f8ea5946ae0fbec7228bf32
